Fully rework the communication, refactoring
The commands used to be sent using a plaintext protocol. Now there is
a single binary packet that can be used to both control all the
functions and get back all the info. Raw LANC data can be passed
through. This makes a lot of things much simpler, as well as adds
predictable feedback from the device.